+++ NEWS.md
@@ -2,6 +2,12 @@
 
 # development version
 
+- Add Izat 2025 zaleplon ([doi:10.1002/psp4.13255](https://doi.org/10.1002/psp4.13255)) -- healthy adult women and men, intravenous.
+
+- Add Izat 2025 ziprasidone ([doi:10.1002/psp4.13255](https://doi.org/10.1002/psp4.13255)) -- healthy adult men, intravenous.
+
+- Add Izat 2025 zoniporide ([doi:10.1002/psp4.13255](https://doi.org/10.1002/psp4.13255)) -- healthy adult men, intravenous.
+
 - Add Moein 2024 apitolisib translational PK/PD models ([doi:10.1007/s40268-024-00459-5](https://doi.org/10.1007/s40268-024-00459-5)) -- mouse (786-O renal cell adenocarcinoma xenograft) and adults with advanced solid tumors or non-Hodgkin's lymphoma.
 
 - Add Decrane 2023 oxyfluorfen ([doi:10.1016/j.crtox.2023.100138](https://doi.org/10.1016/j.crtox.2023.100138)) -- rat (Sprague-Dawley) and extrapolated human.